pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "duino-api"
version = "1.0.0"
description = "Cloud-agnostic hyperscale AI inference platform — Duino API"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }

dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.7.0",
    "httpx>=0.27.0",
    "redis>=5.0.4",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "typer>=0.12.3",
    "rich>=13.7.1",
    "python-dotenv>=1.0.1",
    "prometheus-client>=0.20.0",
    "opentelemetry-api>=1.24.0",
    "opentelemetry-sdk>=1.24.0",
    "opentelemetry-instrumentation-fastapi>=0.45b0",
    "pyngrok>=7.2.0",
    "aiofiles>=23.2.1",
    "psutil>=5.9.8",
]

[project.optional-dependencies]
inference = [
    "torch>=2.3.0",
    "transformers>=4.41.0",
    "accelerate>=0.30.0",
    "bitsandbytes>=0.43.1",
    "vllm>=0.4.2; sys_platform == 'linux'",
    "sentencepiece>=0.2.0",
    "einops>=0.8.0",
]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.7",
    "httpx>=0.27.0",
    "ruff>=0.4.4",
]

[project.scripts]
duinobot = "duino_api.cli.main:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["duino_api*"]

[tool.ruff]
line-length = 100
target-version = "py310"
